refactoring and fixed bug in node runnability computation
