+++ modules/executors/utils/scraper_parser.py
